fix(bar): SJRA-1620 inverse chart type and fix issues
