+++ frontend/components/base/charts/bar-charts/horizontal-bar-chart.tsx
@@ -10,34 +10,38 @@ import BarRectangle from '../shapes/bar-rectangle';
 import { BarChartProps } from '../type';
 
 /**
- * Horizontal bar chart with a single series, one bar per category.
+ * Vertical bar chart with a single series, one bar per category.
  *
  * Value (number) must always use "count" key
  *
- *     │  ▓▒
- *     │  ▓▒
- *     │  ▓▒          ▓▒
- *     │  ▓▒    ▓▒    ▓▒
- *     │  ▓▒ ▓▒ ▓▒ ▓▒ ▓▒ ▓▒
+ *     │ Lbl  ▓▒▓▒▓▒▓▒▓▒▓▒
+ *     │ Lbl2 ▓▒▓▒▓▒▓▒
+ *     │ Lbl3 ▓▒▓▒▓▒▓▒▓▒
+ *     │ Lbl4 ▓▒▓▒
+ *     │ Lbl5 ▓▒▓▒▓▒
  *     └─────────────────────
- *        Lbl   Lbl2  Lbl3
- *
  */
-function HorizontalBarChart<T extends object>({ axis, colorblindMode = true, data, tooltip }: BarChartProps<T>) {
+function HorizontalBarChart<T extends object>({
+  axis,
+  data,
+  colorblindMode = true,
+  onClick,
+  tooltip,
+}: BarChartProps<T>) {
   const id = useId();
   const chartData = useChartPalette(data);
   const chartConfig = useBarChartConfig(axis);
 
   return (
     <ChartContainer config={chartConfig}>
-      <RechartBarChart accessibilityLayer data={chartData} layout="horizontal" margin={{ bottom: 12, left: 6 }}>
+      <RechartBarChart accessibilityLayer data={chartData} layout="vertical" margin={{ bottom: 12, left: 6 }}>
         <ChartPalette id={id} data={chartData} colorblindMode={colorblindMode} />
 
         <CartesianGrid />
 
-        <ChartAxis axis={axis} layout="horizontal" />
+        <ChartAxis axis={axis} layout="vertical" />
 
-        <BarRectangle dataKey="count" id={id} layout="horizontal" />
+        <BarRectangle dataKey="count" id={id} layout="vertical" onClick={onClick} />
 
         <ChartTooltip
           isAnimationActive={false}
