+++ frontend/components/base/charts/bar-charts/vertical-bar-chart.tsx
@@ -14,28 +14,30 @@ import { BarChartProps } from '../type';
  *
  * Value (number) must always use "count" key
  *
- *     │ Lbl  ▓▒▓▒▓▒▓▒▓▒▓▒
- *     │ Lbl2 ▓▒▓▒▓▒▓▒
- *     │ Lbl3 ▓▒▓▒▓▒▓▒▓▒
- *     │ Lbl4 ▓▒▓▒
- *     │ Lbl5 ▓▒▓▒▓▒
+ *     │  ▓▒
+ *     │  ▓▒
+ *     │  ▓▒          ▓▒
+ *     │  ▓▒    ▓▒    ▓▒
+ *     │  ▓▒ ▓▒ ▓▒ ▓▒ ▓▒ ▓▒
  *     └─────────────────────
+ *        Lbl   Lbl2  Lbl3
+ *
  */
-function VerticalBarChart<T extends object>({ axis, data, colorblindMode = true, onClick, tooltip }: BarChartProps<T>) {
+function VerticalBarChart<T extends object>({ axis, colorblindMode = true, data, onClick, tooltip }: BarChartProps<T>) {
   const id = useId();
   const chartData = useChartPalette(data);
   const chartConfig = useBarChartConfig(axis);
 
   return (
     <ChartContainer config={chartConfig}>
-      <RechartBarChart accessibilityLayer data={chartData} layout="vertical" margin={{ bottom: 12, left: 6 }}>
+      <RechartBarChart accessibilityLayer data={chartData} layout="horizontal" margin={{ bottom: 12, left: 6 }}>
         <ChartPalette id={id} data={chartData} colorblindMode={colorblindMode} />
 
         <CartesianGrid />
 
-        <ChartAxis axis={axis} layout="vertical" />
+        <ChartAxis axis={axis} layout="horizontal" />
 
-        <BarRectangle dataKey="count" id={id} layout="vertical" onClick={onClick} />
+        <BarRectangle dataKey="count" id={id} layout="horizontal" onClick={onClick} />
 
         <ChartTooltip
           isAnimationActive={false}
