+++ frontend/components/stories/charts/horizontal-bar-chart.stories.tsx
@@ -1,69 +1,121 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
+import HorizontalBarChart from '@/components/base/charts/bar-charts/horizontal-bar-chart';
+import ChartPalettePreview from '@/components/base/charts/palettes/chart-palette-preview';
+import { ChartTooltipPayload } from '@/components/base/charts/type';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/base/shadcn/card';
 
 import { StorySection } from '../story-section';
-import { ageAtFirstEngamentIncludeData, ageAtFirstEngamentKFData } from './data';
-import HorizontalBarChart from '@/components/base/charts/bar-charts/horizontal-bar-chart';
-import GroupedHorizontalBarChart from '@/components/base/charts/bar-charts/grouped-horizontal-bar-chart';
-import { ChartTooltipPayload } from '@/components/base/charts/type';
-import ChartPalettePreview from '@/components/base/charts/palettes/chart-palette-preview';
 
-const ageAtFirstEngagementFKProps = {
+import { biospecimensData, filesByDataTypesData, hpoData, mondoData } from './data';
+
+const hpoProps = {
   axis: {
     x: {
-      dataKey: 'age',
-      label: 'Age at First Patient Engagement (years)',
-    },
-    y: {
       dataKey: 'count',
       label: '# of participants',
     },
+    y: {
+      dataKey: 'hpo_id',
+      width: 210,
+      tickFormatter: (value: string) => value.replace(/\(HP:\d+\)/g, ''),
+      label: 'Diagnosis (HPO)',
+    },
   },
-  data: ageAtFirstEngamentKFData,
-  tooltip: (payload: ChartTooltipPayload) => {
-    return (
-      <div className="flex gap-2 items-center">
+  data: hpoData,
+  onClick: (data: any) => {
+    console.warn('data', data);
+  },
+  tooltip: (payload: ChartTooltipPayload) => (
+    <div className="flex gap-2">
+      <div className="m-1">
         <ChartPalettePreview patternIndex={payload.patternIndex} />
-        Participants {payload.count}
       </div>
-    );
-  },
+      <div className="flex-1">
+        {payload.hpo_id}
+        <div>Participants with this exact term: {payload.count}</div>
+        <div>Participants including descendant term: {payload.countWithDescendant}</div>
+      </div>
+    </div>
+  ),
 };
-
-const ageAtFirstEngagementIncludeProps = {
+const mondoProps = {
   axis: {
     x: {
-      dataKey: 'age',
-      label: 'Age at First Patient Engagement (years)',
-    },
-    y: {
-      dataKey: 'trisomy',
+      dataKey: 'count',
       label: '# of participants',
     },
+    y: {
+      dataKey: 'mondo_id',
+      width: 180,
+      tickFormatter: (value: string) => value.replace(/\(MONDO:\d+\)/g, ''),
+      label: 'Diagnosis (MONDO)',
+    },
   },
-  bars: ['trisomy', 'disomy'],
-  data: ageAtFirstEngamentIncludeData,
-  tooltip: (payload: ChartTooltipPayload) => {
-    return (
-      <div>
-        <div className="flex gap-2 items-center">
-          <ChartPalettePreview patternIndex={0} />
-          Participants: {payload.trisomy}
-        </div>
-        <div className="flex gap-2 items-center">
-          <ChartPalettePreview patternIndex={1} />
-          Participants: {payload.disomy}
-        </div>
+  data: mondoData,
+  onClick: (data: any) => {
+    console.warn('data', data);
+  },
+  tooltip: (payload: ChartTooltipPayload) => (
+    <div className="flex gap-2">
+      <div className="m-1">
+        <ChartPalettePreview patternIndex={payload.patternIndex} />
       </div>
-    );
+      <div className="flex-1">
+        {payload.mondo_id}
+        <div>Participants with this exact term: {payload.count}</div>
+        <div>Participants including descendant term: {payload.countWithDescendant}</div>
+      </div>
+    </div>
+  ),
+};
+
+const filesByDataTypesProps = {
+  axis: {
+    x: {
+      dataKey: 'count',
+      label: '# of files',
+    },
+    y: {
+      dataKey: 'key',
+      width: 180,
+      label: 'Data Types',
+    },
   },
+  data: filesByDataTypesData.slice(0, 10),
+  tooltip: (payload: ChartTooltipPayload) => (
+    <div className="flex gap-2 items-center">
+      <ChartPalettePreview patternIndex={payload.patternIndex} />
+      {payload.key} <span className="font-bold">{payload.count}</span>
+    </div>
+  ),
+};
+
+const biospecimensProps = {
+  axis: {
+    x: {
+      dataKey: 'count',
+      label: '# of biospecimens',
+    },
+    y: {
+      dataKey: 'key',
+      width: 140,
+      label: 'Sample Types',
+    },
+  },
+  data: biospecimensData.slice(0, 10),
+  tooltip: (payload: ChartTooltipPayload) => (
+    <div className="flex gap-2 items-center">
+      <ChartPalettePreview patternIndex={payload.patternIndex} />
+      {payload.key} <span className="font-bold">{payload.count}</span>
+    </div>
+  ),
 };
 
 const meta = {
   title: 'Components/Charts/Horizontal Bar Chart',
   component: HorizontalBarChart,
-  args: ageAtFirstEngagementFKProps,
+  args: hpoProps,
 } satisfies Meta<typeof HorizontalBarChart>;
 
 export default meta;
@@ -71,49 +123,83 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
-  render: () => {
-    return (
-      <StorySection title="Horizontal Bar chart">
-        <div className="w-full flex gap-6">
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Age At First Engagement</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <HorizontalBarChart {...ageAtFirstEngagementFKProps} />
-            </CardContent>
-          </Card>
+  render: () => (
+    <StorySection title="Vertical Bar chart">
+      <div className="w-full flex gap-6">
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>HPO (Clickable)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...hpoProps} />
+          </CardContent>
+        </Card>
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Mondo (Clickable)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...mondoProps} />
+          </CardContent>
+        </Card>
+      </div>
 
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Age At First Engagement (Grouped)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <GroupedHorizontalBarChart {...ageAtFirstEngagementIncludeProps} />
-            </CardContent>
-          </Card>
-        </div>
+      <div className="w-full flex gap-6">
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>HPO (colorblindMode off) (clickable)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...hpoProps} colorblindMode={false} />
+          </CardContent>
+        </Card>
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Mondo (colorblindMode off) (clickable)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...mondoProps} colorblindMode={false} />
+          </CardContent>
+        </Card>
+      </div>
 
-        <div className="w-full flex gap-6">
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Age At First Engagement (colorblindMode off)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <HorizontalBarChart {...ageAtFirstEngagementFKProps} colorblindMode={false} />
-            </CardContent>
-          </Card>
+      <div className="w-full flex gap-6">
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Files By Data Types</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...filesByDataTypesProps} />
+          </CardContent>
+        </Card>
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Biospecimens</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...biospecimensProps} />
+          </CardContent>
+        </Card>
+      </div>
 
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Age At First Engagement (colorblindMode off) (grouped)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <GroupedHorizontalBarChart {...ageAtFirstEngagementIncludeProps} colorblindMode={false} />
-            </CardContent>
-          </Card>
-        </div>
-      </StorySection>
-    );
-  },
+      <div className="w-full flex gap-6">
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Files By Data Types (colorblindMode off)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...filesByDataTypesProps} colorblindMode={false} />
+          </CardContent>
+        </Card>
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Biospecimens (colorblindMode off)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <HorizontalBarChart {...biospecimensProps} colorblindMode={false} />
+          </CardContent>
+        </Card>
+      </div>
+    </StorySection>
+  ),
 };
