+++ frontend/components/stories/charts/vertical-bar-chart.stories.tsx
@@ -1,132 +1,72 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
+import GroupedVerticalBarChart from '@/components/base/charts/bar-charts/grouped-vertical-bar-chart';
 import VerticalBarChart from '@/components/base/charts/bar-charts/vertical-bar-chart';
+import ChartPalettePreview from '@/components/base/charts/palettes/chart-palette-preview';
+import { ChartTooltipPayload } from '@/components/base/charts/type';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/base/shadcn/card';
 
 import { StorySection } from '../story-section';
-import { biospecimensData, filesByDataTypesData, hpoData, mondoData } from './data';
-import { ChartTooltipPayload } from '@/components/base/charts/type';
-import ChartPalettePreview from '@/components/base/charts/palettes/chart-palette-preview';
 
-const hpoProps = {
+import { ageAtFirstEngamentIncludeData, ageAtFirstEngamentKFData } from './data';
+
+const ageAtFirstEngagementFKProps = {
   axis: {
     x: {
+      dataKey: 'age',
+      label: 'Age at First Patient Engagement (years)',
+    },
+    y: {
       dataKey: 'count',
       label: '# of participants',
     },
-    y: {
-      dataKey: 'hpo_id',
-      width: 210,
-      tickFormatter: (value: string) => {
-        return value.replace(/\(HP:\d+\)/g, '');
-      },
-      label: 'Diagnosis (HPO)',
-    },
   },
-  data: hpoData,
+  data: ageAtFirstEngamentKFData,
   onClick: (data: any) => {
-    console.log('data', data);
+    console.warn('data', data);
   },
-  tooltip: (payload: ChartTooltipPayload) => {
-    return (
-      <div className="flex gap-2">
-        <div className="m-1">
-          <ChartPalettePreview patternIndex={payload.patternIndex} />
-        </div>
-        <div className="flex-1">
-          {payload.hpo_id}
-          <div>Participants with this exact term: {payload.count}</div>
-          <div>Participants including descendant term: {payload.countWithDescendant}</div>
-        </div>
-      </div>
-    );
-  },
+  tooltip: (payload: ChartTooltipPayload) => (
+    <div className="flex gap-2 items-center">
+      <ChartPalettePreview patternIndex={payload.patternIndex} />
+      Participants {payload.count}
+    </div>
+  ),
 };
-const mondoProps = {
+
+const ageAtFirstEngagementIncludeProps = {
   axis: {
     x: {
-      dataKey: 'count',
+      dataKey: 'age',
+      label: 'Age at First Patient Engagement (years)',
+    },
+    y: {
+      dataKey: 'trisomy',
       label: '# of participants',
     },
-    y: {
-      dataKey: 'mondo_id',
-      width: 180,
-      tickFormatter: (value: string) => {
-        return value.replace(/\(MONDO:\d+\)/g, '');
-      },
-      label: 'Diagnosis (MONDO)',
-    },
   },
-  data: mondoData,
+  bars: ['trisomy', 'disomy'],
+  data: ageAtFirstEngamentIncludeData,
   onClick: (data: any) => {
-    console.log('data', data);
+    console.warn('data', data);
   },
-  tooltip: (payload: ChartTooltipPayload) => {
-    return (
-      <div className="flex gap-2">
-        <div className="m-1">
-          <ChartPalettePreview patternIndex={payload.patternIndex} />
-        </div>
-        <div className="flex-1">
-          {payload.mondo_id}
-          <div>Participants with this exact term: {payload.count}</div>
-          <div>Participants including descendant term: {payload.countWithDescendant}</div>
-        </div>
-      </div>
-    );
-  },
-};
-
-const filesByDataTypesProps = {
-  axis: {
-    x: {
-      dataKey: 'count',
-      label: '# of files',
-    },
-    y: {
-      dataKey: 'key',
-      width: 180,
-      label: 'Data Types',
-    },
-  },
-  data: filesByDataTypesData.slice(0, 10),
-  tooltip: (payload: ChartTooltipPayload) => {
-    return (
+  tooltip: (payload: ChartTooltipPayload) => (
+    <div>
       <div className="flex gap-2 items-center">
-        <ChartPalettePreview patternIndex={payload.patternIndex} />
-        {payload.key} <span className="font-bold">{payload.count}</span>
+        <ChartPalettePreview patternIndex={0} />
+        Participants: {payload.trisomy}
       </div>
-    );
-  },
-};
-
-const biospecimensProps = {
-  axis: {
-    x: {
-      dataKey: 'count',
-      label: '# of biospecimens',
-    },
-    y: {
-      dataKey: 'key',
-      width: 140,
-      label: 'Sample Types',
-    },
-  },
-  data: biospecimensData.slice(0, 10),
-  tooltip: (payload: ChartTooltipPayload) => {
-    return (
       <div className="flex gap-2 items-center">
-        <ChartPalettePreview patternIndex={payload.patternIndex} />
-        {payload.key} <span className="font-bold">{payload.count}</span>
+        <ChartPalettePreview patternIndex={1} />
+        Participants: {payload.disomy}
       </div>
-    );
-  },
+    </div>
+  ),
 };
 
 const meta = {
   title: 'Components/Charts/Vertical Bar Chart',
   component: VerticalBarChart,
-  args: hpoProps,
+  args: ageAtFirstEngagementFKProps,
 } satisfies Meta<typeof VerticalBarChart>;
 
 export default meta;
@@ -134,85 +74,47 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
-  render: () => {
-    return (
-      <StorySection title="Vertical Bar chart">
-        <div className="w-full flex gap-6">
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>HPO (Clickable)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...hpoProps} />
-            </CardContent>
-          </Card>
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Mondo (Clickable)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...mondoProps} />
-            </CardContent>
-          </Card>
-        </div>
+  render: () => (
+    <StorySection title="Horizontal Bar chart">
+      <div className="w-full flex gap-6">
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Age At First Engagement</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <VerticalBarChart {...ageAtFirstEngagementFKProps} />
+          </CardContent>
+        </Card>
 
-        <div className="w-full flex gap-6">
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>HPO (colorblindMode off) (clickable)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...hpoProps} colorblindMode={false} />
-            </CardContent>
-          </Card>
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Mondo (colorblindMode off) (clickable)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...mondoProps} colorblindMode={false} />
-            </CardContent>
-          </Card>
-        </div>
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Age At First Engagement (Grouped)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <GroupedVerticalBarChart {...ageAtFirstEngagementIncludeProps} />
+          </CardContent>
+        </Card>
+      </div>
 
-        <div className="w-full flex gap-6">
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Files By Data Types</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...filesByDataTypesProps} />
-            </CardContent>
-          </Card>
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Biospecimens</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...biospecimensProps} />
-            </CardContent>
-          </Card>
-        </div>
+      <div className="w-full flex gap-6">
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Age At First Engagement (colorblindMode off)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <VerticalBarChart {...ageAtFirstEngagementFKProps} colorblindMode={false} />
+          </CardContent>
+        </Card>
 
-        <div className="w-full flex gap-6">
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Files By Data Types (colorblindMode off)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...filesByDataTypesProps} colorblindMode={false} />
-            </CardContent>
-          </Card>
-          <Card className={`w-full`}>
-            <CardHeader>
-              <CardTitle>Biospecimens (colorblindMode off)</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <VerticalBarChart {...biospecimensProps} colorblindMode={false} />
-            </CardContent>
-          </Card>
-        </div>
-      </StorySection>
-    );
-  },
+        <Card className={`w-full`}>
+          <CardHeader>
+            <CardTitle>Age At First Engagement (colorblindMode off) (grouped)</CardTitle>
+          </CardHeader>
+          <CardContent>
+            <GroupedVerticalBarChart {...ageAtFirstEngagementIncludeProps} colorblindMode={false} />
+          </CardContent>
+        </Card>
+      </div>
+    </StorySection>
+  ),
 };
